Extract overloads to pyi file
