+++ awswrangler/dynamodb/_read.py
@@ -10,13 +10,11 @@ from typing import (
     Dict,
     Iterator,
     List,
-    Literal,
     Optional,
     Sequence,
     TypeVar,
     Union,
     cast,
-    overload,
 )
 
 import boto3
@@ -44,38 +42,6 @@ _ItemsListType = List[Dict[str, Any]]
 def _read_chunked(iterator: Iterator[Dict[str, Any]]) -> Iterator[pd.DataFrame]:
     for item in iterator:
         yield pd.DataFrame(item)
-
-
-@overload
-def read_partiql_query(
-    query: str,
-    parameters: Optional[List[Any]] = ...,
-    chunked: Literal[False] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-) -> pd.DataFrame:
-    ...
-
-
-@overload
-def read_partiql_query(
-    query: str,
-    *,
-    parameters: Optional[List[Any]] = ...,
-    chunked: Literal[True],
-    boto3_session: Optional[boto3.Session] = ...,
-) -> Iterator[pd.DataFrame]:
-    ...
-
-
-@overload
-def read_partiql_query(
-    query: str,
-    *,
-    parameters: Optional[List[Any]] = ...,
-    chunked: bool,
-    boto3_session: Optional[boto3.Session] = ...,
-) -> Union[pd.DataFrame, Iterator[pd.DataFrame]]:
-    ...
 
 
 def read_partiql_query(
@@ -430,221 +396,6 @@ def _read_items(
         )
     else:
         return _convert_items(items=cast(_ItemsListType, items), as_dataframe=as_dataframe, arrow_kwargs=arrow_kwargs)
-
-
-@overload
-def read_items(
-    table_name: str,
-    index_name: Optional[str] = ...,
-    partition_values: Optional[Sequence[Any]] = ...,
-    sort_values: Optional[Sequence[Any]] = ...,
-    filter_expression: Optional[Union[ConditionBase, str]] = ...,
-    key_condition_expression: Optional[Union[ConditionBase, str]] = ...,
-    expression_attribute_names: Optional[Dict[str, str]] = ...,
-    expression_attribute_values: Optional[Dict[str, Any]] = ...,
-    consistent: bool = ...,
-    columns: Optional[Sequence[str]] = ...,
-    allow_full_scan: bool = ...,
-    max_items_evaluated: Optional[int] = ...,
-    as_dataframe: Literal[True] = ...,
-    chunked: Literal[False] = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> pd.DataFrame:
-    ...
-
-
-@overload
-def read_items(
-    table_name: str,
-    *,
-    index_name: Optional[str] = ...,
-    partition_values: Optional[Sequence[Any]] = ...,
-    sort_values: Optional[Sequence[Any]] = ...,
-    filter_expression: Optional[Union[ConditionBase, str]] = ...,
-    key_condition_expression: Optional[Union[ConditionBase, str]] = ...,
-    expression_attribute_names: Optional[Dict[str, str]] = ...,
-    expression_attribute_values: Optional[Dict[str, Any]] = ...,
-    consistent: bool = ...,
-    columns: Optional[Sequence[str]] = ...,
-    allow_full_scan: bool = ...,
-    max_items_evaluated: Optional[int] = ...,
-    as_dataframe: Literal[True] = ...,
-    chunked: Literal[True],
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Iterator[pd.DataFrame]:
-    ...
-
-
-@overload
-def read_items(
-    table_name: str,
-    *,
-    index_name: Optional[str] = ...,
-    partition_values: Optional[Sequence[Any]] = ...,
-    sort_values: Optional[Sequence[Any]] = ...,
-    filter_expression: Optional[Union[ConditionBase, str]] = ...,
-    key_condition_expression: Optional[Union[ConditionBase, str]] = ...,
-    expression_attribute_names: Optional[Dict[str, str]] = ...,
-    expression_attribute_values: Optional[Dict[str, Any]] = ...,
-    consistent: bool = ...,
-    columns: Optional[Sequence[str]] = ...,
-    allow_full_scan: bool = ...,
-    max_items_evaluated: Optional[int] = ...,
-    as_dataframe: Literal[False],
-    chunked: Literal[False] = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> _ItemsListType:
-    ...
-
-
-@overload
-def read_items(
-    table_name: str,
-    *,
-    index_name: Optional[str] = ...,
-    partition_values: Optional[Sequence[Any]] = ...,
-    sort_values: Optional[Sequence[Any]] = ...,
-    filter_expression: Optional[Union[ConditionBase, str]] = ...,
-    key_condition_expression: Optional[Union[ConditionBase, str]] = ...,
-    expression_attribute_names: Optional[Dict[str, str]] = ...,
-    expression_attribute_values: Optional[Dict[str, Any]] = ...,
-    consistent: bool = ...,
-    columns: Optional[Sequence[str]] = ...,
-    allow_full_scan: bool = ...,
-    max_items_evaluated: Optional[int] = ...,
-    as_dataframe: Literal[False],
-    chunked: Literal[True],
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Iterator[_ItemsListType]:
-    ...
-
-
-@overload
-def read_items(
-    table_name: str,
-    *,
-    index_name: Optional[str] = ...,
-    partition_values: Optional[Sequence[Any]] = ...,
-    sort_values: Optional[Sequence[Any]] = ...,
-    filter_expression: Optional[Union[ConditionBase, str]] = ...,
-    key_condition_expression: Optional[Union[ConditionBase, str]] = ...,
-    expression_attribute_names: Optional[Dict[str, str]] = ...,
-    expression_attribute_values: Optional[Dict[str, Any]] = ...,
-    consistent: bool = ...,
-    columns: Optional[Sequence[str]] = ...,
-    allow_full_scan: bool = ...,
-    max_items_evaluated: Optional[int] = ...,
-    as_dataframe: bool,
-    chunked: Literal[False] = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Union[pd.DataFrame, _ItemsListType]:
-    ...
-
-
-@overload
-def read_items(
-    table_name: str,
-    *,
-    index_name: Optional[str] = ...,
-    partition_values: Optional[Sequence[Any]] = ...,
-    sort_values: Optional[Sequence[Any]] = ...,
-    filter_expression: Optional[Union[ConditionBase, str]] = ...,
-    key_condition_expression: Optional[Union[ConditionBase, str]] = ...,
-    expression_attribute_names: Optional[Dict[str, str]] = ...,
-    expression_attribute_values: Optional[Dict[str, Any]] = ...,
-    consistent: bool = ...,
-    columns: Optional[Sequence[str]] = ...,
-    allow_full_scan: bool = ...,
-    max_items_evaluated: Optional[int] = ...,
-    as_dataframe: bool,
-    chunked: Literal[True],
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Union[Iterator[pd.DataFrame], Iterator[_ItemsListType]]:
-    ...
-
-
-@overload
-def read_items(
-    table_name: str,
-    *,
-    index_name: Optional[str] = ...,
-    partition_values: Optional[Sequence[Any]] = ...,
-    sort_values: Optional[Sequence[Any]] = ...,
-    filter_expression: Optional[Union[ConditionBase, str]] = ...,
-    key_condition_expression: Optional[Union[ConditionBase, str]] = ...,
-    expression_attribute_names: Optional[Dict[str, str]] = ...,
-    expression_attribute_values: Optional[Dict[str, Any]] = ...,
-    consistent: bool = ...,
-    columns: Optional[Sequence[str]] = ...,
-    allow_full_scan: bool = ...,
-    max_items_evaluated: Optional[int] = ...,
-    as_dataframe: Literal[True],
-    chunked: bool,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Union[pd.DataFrame, Iterator[pd.DataFrame]]:
-    ...
-
-
-@overload
-def read_items(
-    table_name: str,
-    *,
-    index_name: Optional[str] = ...,
-    partition_values: Optional[Sequence[Any]] = ...,
-    sort_values: Optional[Sequence[Any]] = ...,
-    filter_expression: Optional[Union[ConditionBase, str]] = ...,
-    key_condition_expression: Optional[Union[ConditionBase, str]] = ...,
-    expression_attribute_names: Optional[Dict[str, str]] = ...,
-    expression_attribute_values: Optional[Dict[str, Any]] = ...,
-    consistent: bool = ...,
-    columns: Optional[Sequence[str]] = ...,
-    allow_full_scan: bool = ...,
-    max_items_evaluated: Optional[int] = ...,
-    as_dataframe: Literal[False],
-    chunked: bool,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Union[_ItemsListType, Iterator[_ItemsListType]]:
-    ...
-
-
-@overload
-def read_items(
-    table_name: str,
-    *,
-    index_name: Optional[str] = ...,
-    partition_values: Optional[Sequence[Any]] = ...,
-    sort_values: Optional[Sequence[Any]] = ...,
-    filter_expression: Optional[Union[ConditionBase, str]] = ...,
-    key_condition_expression: Optional[Union[ConditionBase, str]] = ...,
-    expression_attribute_names: Optional[Dict[str, str]] = ...,
-    expression_attribute_values: Optional[Dict[str, Any]] = ...,
-    consistent: bool = ...,
-    columns: Optional[Sequence[str]] = ...,
-    allow_full_scan: bool = ...,
-    max_items_evaluated: Optional[int] = ...,
-    as_dataframe: bool,
-    chunked: bool,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Union[pd.DataFrame, Iterator[pd.DataFrame], _ItemsListType, Iterator[_ItemsListType]]:
-    ...
 
 
 @_utils.validate_distributed_kwargs(
